requirements.txt
opencv-python>=4.7.0
pyautogui>=0.9.53
pydirectinput>=1.0.4
numpy>=1.24.0
pillow>=9.5.0
psutil>=5.9.0
pywin32>=305
pynput>=1.7.6
keyboard>=0.13.5
steam>=1.4.4
websockets>=11.0.3
asyncio>=3.4.3